Break out logs appender API
