+++ api/logs-appender/README.md
@@ -1,0 +1,6 @@
+# OpenTelemetry Experimental Logs Appender API
+
+[![Javadocs][javadoc-image]][javadoc-url]
+
+[javadoc-image]: https://www.javadoc.io/badge/io.opentelemetry/opentelemetry-api-logs-appender.svg
+[javadoc-url]: https://www.javadoc.io/doc/io.opentelemetry/opentelemetry-api-logs-appender
