+++ api/logs-appender/src/main/java/io/opentelemetry/api/logs/LogEmitter.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.sdk.logs;
+package io.opentelemetry.api.logs;
 
 import javax.annotation.concurrent.ThreadSafe;
 
@@ -11,7 +11,7 @@ import javax.annotation.concurrent.ThreadSafe;
  * A {@link LogEmitter} is the entry point into a log pipeline.
  *
  * <p>Obtain a log builder via {@link #logBuilder()}, add properties using the setters, and emit it
- * to downstream {@link LogProcessor}(s) via {@link LogBuilder#emit()}.
+ * to the pipeline via {@link LogBuilder#emit()}.
  */
 @ThreadSafe
 public interface LogEmitter {
@@ -19,8 +19,8 @@ public interface LogEmitter {
   /**
    * Return a {@link LogBuilder} to emit a log.
    *
-   * <p>Build the log using the {@link LogBuilder} setters, and emit it to downstream {@link
-   * LogProcessor}(s) via {@link LogBuilder#emit()}.
+   * <p>Build the log using the {@link LogBuilder} setters, and emit it to the pipeline via {@link
+   * LogBuilder#emit()}.
    */
   LogBuilder logBuilder();
 }
