+++ api/logs-appender/src/main/java/io/opentelemetry/api/logs/package-info.java
@@ -1,0 +1,10 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/** The API for log appenders to interact with OpenTelemetry. */
+@ParametersAreNonnullByDefault
+package io.opentelemetry.api.logs;
+
+import javax.annotation.ParametersAreNonnullByDefault;
