+++ sdk/logs/README.md
@@ -1,11 +1,6 @@
-# OpenTelemetry Experimental Logging Support
+# OpenTelemetry Experimental Logs SDK
 
 [![Javadocs][javadoc-image]][javadoc-url]
 
-This project contains experimental support for transport of logs via OpenTelemetry. The API
-presented is intended for the use of logging library adapters to enable resource and request
-correlation with other observability signals and transport of logs through the OpenTelemetry
-collector.
-
-[javadoc-image]: https://www.javadoc.io/badge/io.opentelemetry/opentelemetry-sdk-extension-logging.svg
-[javadoc-url]: https://www.javadoc.io/doc/io.opentelemetry/opentelemetry-sdk-extension-logging
+[javadoc-image]: https://www.javadoc.io/badge/io.opentelemetry/opentelemetry-sdk-logs.svg
+[javadoc-url]: https://www.javadoc.io/doc/io.opentelemetry/opentelemetry-sdk-logs
