+++ sdk/logs/src/main/java/io/opentelemetry/sdk/logs/LogProcessor.java
@@ -5,6 +5,7 @@
 
 package io.opentelemetry.sdk.logs;
 
+import io.opentelemetry.api.logs.LogEmitter;
 import io.opentelemetry.sdk.common.CompletableResultCode;
 import io.opentelemetry.sdk.logs.data.LogData;
 import java.io.Closeable;
