+++ sdk/logs/src/main/java/io/opentelemetry/sdk/logs/SdkLogEmitterProvider.java
@@ -5,6 +5,9 @@
 
 package io.opentelemetry.sdk.logs;
 
+import io.opentelemetry.api.logs.LogEmitter;
+import io.opentelemetry.api.logs.LogEmitterBuilder;
+import io.opentelemetry.api.logs.LogEmitterProvider;
 import io.opentelemetry.sdk.common.Clock;
 import io.opentelemetry.sdk.common.CompletableResultCode;
 import io.opentelemetry.sdk.internal.ComponentRegistry;
@@ -16,7 +19,7 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 
 /** SDK registry for creating {@link LogEmitter}s. */
-public final class SdkLogEmitterProvider implements Closeable {
+public final class SdkLogEmitterProvider implements LogEmitterProvider, Closeable {
 
   static final String DEFAULT_EMITTER_NAME = "unknown";
   private static final Logger LOGGER = Logger.getLogger(SdkLogEmitterProvider.class.getName());
@@ -47,6 +50,7 @@ public final class SdkLogEmitterProvider implements Closeable {
    * @param instrumentationName the name of the instrumentation library
    * @return a log emitter builder instance
    */
+  @Override
   public LogEmitterBuilder logEmitterBuilder(String instrumentationName) {
     if (instrumentationName == null || instrumentationName.isEmpty()) {
       LOGGER.fine("LogEmitter requested without instrumentation name.");
