+++ sdk/logs/src/main/java/io/opentelemetry/sdk/logs/data/LogDataImpl.java
@@ -7,6 +7,7 @@ package io.opentelemetry.sdk.logs.data;
 
 import com.google.auto.value.AutoValue;
 import io.opentelemetry.api.common.Attributes;
+import io.opentelemetry.api.logs.Severity;
 import io.opentelemetry.api.trace.SpanContext;
 import io.opentelemetry.sdk.common.InstrumentationLibraryInfo;
 import io.opentelemetry.sdk.resources.Resource;
