+++ sdk/logs/src/test/java/io/opentelemetry/sdk/logs/export/BatchLogProcessorTest.java
@@ -10,12 +10,12 @@ import static org.assertj.core.api.Assertions.fail;
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 import static org.awaitility.Awaitility.await;
 
+import io.opentelemetry.api.logs.LogEmitter;
+import io.opentelemetry.api.logs.Severity;
 import io.opentelemetry.sdk.common.CompletableResultCode;
-import io.opentelemetry.sdk.logs.LogEmitter;
 import io.opentelemetry.sdk.logs.LogProcessor;
 import io.opentelemetry.sdk.logs.SdkLogEmitterProvider;
 import io.opentelemetry.sdk.logs.data.LogData;
-import io.opentelemetry.sdk.logs.data.Severity;
 import io.opentelemetry.sdk.logs.util.TestLogExporter;
 import java.time.Duration;
 import java.util.concurrent.TimeUnit;
