+++ sdk/logs/src/test/java/io/opentelemetry/sdk/logs/util/TestUtil.java
@@ -6,10 +6,10 @@
 package io.opentelemetry.sdk.logs.util;
 
 import io.opentelemetry.api.common.Attributes;
+import io.opentelemetry.api.logs.Severity;
 import io.opentelemetry.sdk.common.InstrumentationLibraryInfo;
 import io.opentelemetry.sdk.logs.data.LogData;
 import io.opentelemetry.sdk.logs.data.LogDataBuilder;
-import io.opentelemetry.sdk.logs.data.Severity;
 import io.opentelemetry.sdk.resources.Resource;
 import java.util.concurrent.TimeUnit;
